+++ src/main/java/edu/depaul/armada/model/DashboardContainer.java
@@ -22,6 +22,6 @@ public class DashboardContainer implements Serializable {
 	public long memTotal;
 	public long diskUsed;
 	public long diskTotal;
-	public int cpuUsed;
-	public int cpuTotal;
+	public long cpuUsed;
+	public long cpuTotal;
 }
