+++ src/main/java/edu/depaul/armada/model/DashboardContainerLog.java
@@ -16,7 +16,7 @@ public class DashboardContainerLog implements Serializable {
 	public long memTotal;
 	public long diskUsed;
 	public long diskTotal;
-	public int cpuUsed;
-	public int cpuTotal;
+	public long cpuUsed;
+	public long cpuTotal;
 
 }
